Allow plugins to be supplied in directories
Plugins can now either be a single file (e.g.: `plugins/help.py`) or a directory (e.g.: `plugins/help/help.py`).
If plugins are supplied as a directory, config and stored data will be stored and loaded from the directory, so all relevant files can live inside the plugin's directory.

Attention: as of this commit, loading stored data for plugins moved from single-file to directory can *not* be loaded.

This prepares for the completion of #17, as it puts all loading logic into pluginloader.py and thus does not require any addition files in the plugins-subdirectory anymore.