Breaking Change: Store plugin_data as json-files instead of pickle
From this commit on, plugin_data will be stored as json-files as created by jsonpickle.
This has the following implications:
- jsonpickle added to requirements.txt - make sure to update your requirements (e.g. `pip install -r requirements.txt`)
- Dicts stored via plugin.store_data() can no longer use `int` as keys. Upon loading them from json, they will be loaded as `str`.
- Existing plugin_data will automatically converted to json on startup. After verifying correct execution of plugins, the old .pkl-file can be deleted.

Changing to json should make storing data future-proof and allows for easier readability (for humans and machines).
This allows to move existing data between single-file and directory-based plugins in the future, as needed for #17.