Join DMs by accounts that are not a botmaster.
Since verification of room type is not reliable before actually joining, verify if it is actually a DM after joining. Should the room prove not to be a DM, leave it immediately.
Reject invites that are deemed unauthorised right away instead of not reacting to them.
Send only one join event instead of three, there seems to be no need to actually send multiple joins.

fixes #25