Expandable message: Additional content in messages and notices
As this is currently not in the matrix spec, it's using [<details>: The Details disclosure element](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details). As such, it's not being rendered properly on all clients.
Known to work:
- Element Web

Known not to work:
- Element Android
- Element iOS

Since the content is still visible on clients not rendering the expandable message correctly, not much breakage is expected when using expandable messages.